refactor(meta): move some "impl FromToProto".
