+++ src/meta/proto-conv/src/file_format_from_to_protobuf_impl.rs
@@ -17,9 +17,13 @@
 
 use common_meta_app as mt;
 use common_protos::pb;
+use num::FromPrimitive;
 
+use crate::reader_check_msg;
 use crate::FromToProto;
 use crate::Incompatible;
+use crate::MIN_READER_VER;
+use crate::VER;
 
 impl FromToProto for mt::principal::StageFileFormatType {
     type PB = pb::StageFileFormatType;
@@ -97,5 +101,106 @@ impl FromToProto for mt::principal::StageFileCompression {
             mt::principal::StageFileCompression::None => Ok(pb::StageFileCompression::None),
             mt::principal::StageFileCompression::Xz => Ok(pb::StageFileCompression::Xz),
         }
+    }
+}
+
+impl FromToProto for mt::principal::FileFormatOptions {
+    type PB = pb::FileFormatOptions;
+    fn get_pb_ver(p: &Self::PB) -> u64 {
+        p.ver
+    }
+    fn from_pb(p: pb::FileFormatOptions) -> Result<Self, Incompatible>
+    where Self: Sized {
+        reader_check_msg(p.ver, p.min_reader_ver)?;
+
+        let format = mt::principal::StageFileFormatType::from_pb(
+            FromPrimitive::from_i32(p.format).ok_or_else(|| Incompatible {
+                reason: format!("invalid StageFileFormatType: {}", p.format),
+            })?,
+        )?;
+
+        let compression = mt::principal::StageFileCompression::from_pb(
+            FromPrimitive::from_i32(p.compression).ok_or_else(|| Incompatible {
+                reason: format!("invalid StageFileCompression: {}", p.compression),
+            })?,
+        )?;
+
+        let nan_display = if p.nan_display.is_empty() {
+            "".to_string()
+        } else {
+            p.nan_display
+        };
+
+        Ok(mt::principal::FileFormatOptions {
+            format,
+            skip_header: p.skip_header,
+            field_delimiter: p.field_delimiter.clone(),
+            record_delimiter: p.record_delimiter,
+            nan_display,
+            escape: p.escape,
+            compression,
+            row_tag: p.row_tag,
+            quote: p.quote,
+            name: None,
+        })
+    }
+
+    fn to_pb(&self) -> Result<pb::FileFormatOptions, Incompatible> {
+        let format = mt::principal::StageFileFormatType::to_pb(&self.format)? as i32;
+        let compression = mt::principal::StageFileCompression::to_pb(&self.compression)? as i32;
+        Ok(pb::FileFormatOptions {
+            ver: VER,
+            min_reader_ver: MIN_READER_VER,
+            format,
+            skip_header: self.skip_header,
+            field_delimiter: self.field_delimiter.clone(),
+            record_delimiter: self.record_delimiter.clone(),
+            nan_display: self.nan_display.clone(),
+            compression,
+            row_tag: self.row_tag.clone(),
+            escape: self.escape.clone(),
+            quote: self.quote.clone(),
+        })
+    }
+}
+
+impl FromToProto for mt::principal::UserDefinedFileFormat {
+    type PB = pb::UserDefinedFileFormat;
+    fn get_pb_ver(p: &Self::PB) -> u64 {
+        p.ver
+    }
+    fn from_pb(p: pb::UserDefinedFileFormat) -> Result<Self, Incompatible>
+    where Self: Sized {
+        reader_check_msg(p.ver, p.min_reader_ver)?;
+
+        let file_format_options =
+            mt::principal::FileFormatOptions::from_pb(p.file_format_options.ok_or_else(|| {
+                Incompatible {
+                    reason: "StageInfo.file_format_options cannot be None".to_string(),
+                }
+            })?)?;
+        let creator =
+            mt::principal::UserIdentity::from_pb(p.creator.ok_or_else(|| Incompatible {
+                reason: "StageInfo.file_format_options cannot be None".to_string(),
+            })?)?;
+
+        Ok(mt::principal::UserDefinedFileFormat {
+            name: p.name,
+            file_format_options,
+            creator,
+        })
+    }
+
+    fn to_pb(&self) -> Result<pb::UserDefinedFileFormat, Incompatible> {
+        let file_format_options =
+            mt::principal::FileFormatOptions::to_pb(&self.file_format_options)?;
+        let creator = mt::principal::UserIdentity::to_pb(&self.creator)?;
+        Ok(pb::UserDefinedFileFormat {
+            ver: VER,
+            min_reader_ver: MIN_READER_VER,
+            name: self.name.clone(),
+            file_format_options: Some(file_format_options),
+            creator: Some(creator),
+        })
     }
 }
